+++ scg-impl/src/main/java/com/dwarfeng/scg/impl/service/telqos/GenerateCommand.java
@@ -1,33 +1,38 @@
 package com.dwarfeng.scg.impl.service.telqos;
 
 import com.dwarfeng.scg.stack.service.GenerateQosService;
+import com.dwarfeng.springtelqos.node.config.TelqosCommand;
 import com.dwarfeng.springtelqos.sdk.command.CliCommand;
 import com.dwarfeng.springtelqos.stack.command.Context;
 import com.dwarfeng.springtelqos.stack.exception.TelqosException;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
-import org.apache.commons.cli.ParseException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Component;
 
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileWriter;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
-@Component
+@TelqosCommand
 public class GenerateCommand extends CliCommand {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(GenerateCommand.class);
-
-    private static final String COMMAND_OPTION_SCG_SETTING_KEY = "s";
-    private static final String COMMAND_OPTION_SIZE = "i";
+    private static final String CMD_OPTION_SCG_SETTING_ID = "id";
+    private static final String CMD_OPTION_SIZE = "s";
+    private static final String CMD_OPTION_FILE = "f";
 
     private static final String IDENTITY = "gen";
-    private static final String DESCRIPTION = "数据记录本地缓存操作";
+    private static final String DESCRIPTION = "流水码生成/导出操作";
 
-    private static final String CMD_LINE_SYNTAX = "gen -" + COMMAND_OPTION_SCG_SETTING_KEY +
-            " [id] " + "-" + COMMAND_OPTION_SIZE + " [size]";
+    private static final String CMD_LINE_SYNTAX = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(CMD_OPTION_SCG_SETTING_ID) + " scg-setting-id [" +
+            CommandUtil.concatOptionPrefix(CMD_OPTION_SIZE) + " size] [" +
+            CommandUtil.concatOptionPrefix(CMD_OPTION_FILE) + " [file-path]]";
+
+    private static final int DEFAULT_SIZE = 5;
+    private static final String DEFAULT_FILE_NAME = "export.csv";
 
     public GenerateCommand(GenerateQosService generateQosService) {
         super(IDENTITY, DESCRIPTION, CMD_LINE_SYNTAX);
@@ -39,32 +44,92 @@ public class GenerateCommand extends CliCommand {
     @Override
     protected List<Option> buildOptions() {
         List<Option> list = new ArrayList<>();
-        list.add(Option.builder(COMMAND_OPTION_SCG_SETTING_KEY).hasArg().desc("流水码生成设置ID").type(String.class).required().build());
-        list.add(Option.builder(COMMAND_OPTION_SIZE).hasArg().desc("流水码生成数量").type(Number.class).required().build());
+        list.add(
+                Option.builder(CMD_OPTION_SCG_SETTING_ID).hasArg().argName("scg-setting-id")
+                        .desc("流水码生成设置 ID").type(String.class).required().build()
+        );
+        list.add(
+                Option.builder(CMD_OPTION_SIZE).hasArg().argName("size").desc("生成数量").type(Number.class)
+                        .build()
+        );
+        list.add(
+                Option.builder(CMD_OPTION_FILE).optionalArg(true).hasArg(true).argName("file-path")
+                        .desc("生成的CSV的路径").type(String.class).build()
+        );
         return list;
     }
 
     @Override
     protected void executeWithCmd(Context context, CommandLine cmd) throws TelqosException {
         try {
-            String scgSettingId = ((String) cmd.getParsedOptionValue(COMMAND_OPTION_SCG_SETTING_KEY));
+            String scgSettingId = ((String) cmd.getParsedOptionValue(CMD_OPTION_SCG_SETTING_ID));
             int size;
-            try {
-                size = ((Number) cmd.getParsedOptionValue(COMMAND_OPTION_SIZE)).intValue();
-            } catch (ParseException e) {
-                LOGGER.warn("解析命令选项时发生异常，异常信息如下", e);
-                context.sendMessage("命令行格式错误，正确的格式为: " + CMD_LINE_SYNTAX);
-                context.sendMessage("请留意选项 " + COMMAND_OPTION_SIZE + " 后接参数的类型应该是数字 ");
-                return;
+            if (cmd.hasOption(CMD_OPTION_SIZE)) {
+                Number sizeNumber = (Number) cmd.getParsedOptionValue(CMD_OPTION_SIZE);
+                if (Objects.isNull(sizeNumber)) {
+                    size = DEFAULT_SIZE;
+                } else {
+                    size = sizeNumber.intValue();
+                }
+            } else {
+                context.sendMessage("请输入需要生成的 ID 的数量:");
+                size = Integer.parseInt(context.receiveMessage());
             }
-            List<String> result = generateQosService.batchGenerate(new StringIdKey(scgSettingId), size);
-            int digits = digits(result.size());
-            for (int i = 0; i < result.size(); i++) {
-                String serialCode = result.get(i);
-                context.sendMessage(String.format("%" + digits + "d: %s", i, serialCode));
+            List<String> serialCodes = generateQosService.batchGenerate(new StringIdKey(scgSettingId), size);
+            if (cmd.hasOption(CMD_OPTION_FILE)) {
+                print2File(serialCodes, context, cmd);
+            } else {
+                print2Screen(serialCodes, context);
             }
         } catch (Exception e) {
             throw new TelqosException(e);
+        }
+    }
+
+    private void print2File(List<String> serialCodes, Context context, CommandLine cmd) throws Exception {
+        File file;
+
+        String optionFileName = (String) cmd.getParsedOptionValue(CMD_OPTION_FILE);
+        if (Objects.nonNull(optionFileName)) {
+            file = new File((optionFileName));
+            if (file.isDirectory()) {
+                file = new File(file, DEFAULT_FILE_NAME);
+            }
+        } else {
+            file = new File(DEFAULT_FILE_NAME);
+        }
+
+        if (file.exists()) {
+            while (true) {
+                context.sendMessage("文件 \"" + file.getName() + "\" 已经存在，是否覆盖？ Y/N");
+                String option = context.receiveMessage().toUpperCase();
+                if (Objects.equals(option, "Y")) {
+                    break;
+                } else if (Objects.equals(option, "N")) {
+                    context.sendMessage("操作已取消！");
+                    return;
+                }
+            }
+        }
+
+        try (
+                FileWriter fw = new FileWriter(file);
+                BufferedWriter bw = new BufferedWriter(fw, 4096)
+        ) {
+            for (String serialCode : serialCodes) {
+                bw.write(serialCode);
+                bw.write(System.lineSeparator());
+            }
+            bw.flush();
+        }
+        context.sendMessage("CSV 文件已导出至 " + file.getAbsolutePath());
+    }
+
+    private void print2Screen(List<String> serialCodes, Context context) throws Exception {
+        int digits = digits(serialCodes.size() - 1);
+        for (int i = 0; i < serialCodes.size(); i++) {
+            String serialCode = serialCodes.get(i);
+            context.sendMessage(String.format("%" + digits + "d: %s", i, serialCode));
         }
     }